+++ frontend/src/app/chat/chat-page/chat-page.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnDestroy, OnInit, ViewChild, ViewEncapsulation, inject } from '@angular/core';
+import { TextFieldModule } from '@angular/cdk/text-field';
+import { Component, ElementRef, OnDestroy, OnInit, ViewChild, ViewEncapsulation, inject } from '@angular/core';
 import { ChatService } from '../chat.service';
 import { AsyncPipe, CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
@@ -20,6 +21,7 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { ChatListComponent } from "../chat-list/chat-list.component";
 import { ChatViewComponent } from "../chat-view/chat-view.component";
 import { SessionService } from '../../shared/session.service';
+import { SpeechSynthesisService } from '../../shared/speech-synthesis.service';
 import { SpeechRecognitionButtonComponent } from "../../shared/speech-recognition-button/speech-recognition-button.component";
 
 
@@ -37,12 +39,13 @@ import { SpeechRecognitionButtonComponent } from "../../shared/speech-recognitio
         MatListModule,
         MatIconModule,
         AsyncPipe,
-        FormsModule, MatInputModule, MatTooltipModule, MatMenuModule, MarkdownPipe, MatProgressSpinnerModule, ChatListComponent, ChatViewComponent, SpeechRecognitionButtonComponent]
+        FormsModule, MatInputModule, MatTooltipModule, MatMenuModule, MarkdownPipe, MatProgressSpinnerModule, ChatListComponent, ChatViewComponent, SpeechRecognitionButtonComponent, TextFieldModule]
 })
 export class ChatPageComponent implements OnInit, OnDestroy {
 
     @ViewChild(MatSidenavContainer) drawerContainer!: MatSidenavContainer;
     @ViewChild('container') chatView!: ChatViewComponent;
+    @ViewChild('messageInput') messageInput!: ElementRef<HTMLTextAreaElement>;
 
     models: string[] = [];
     model!: string;
@@ -50,6 +53,11 @@ export class ChatPageComponent implements OnInit, OnDestroy {
     sessionChanged = false;
     newMessage = '';
     progressSpinner = false;
+    /**
+     * Indicates if the user is speaking and
+     * the response should be spoken too
+     */
+    isSpeech = false;
 
 
     private dataSubscription!: Subscription;
@@ -62,7 +70,7 @@ export class ChatPageComponent implements OnInit, OnDestroy {
         );
     isHandset!: boolean;
 
-    constructor(public authService: AuthService, private config: ConfigService, public chatService: ChatService, public sessionService: SessionService) {
+    constructor(public authService: AuthService, private config: ConfigService, private speechSynthesis: SpeechSynthesisService, public sessionService: SessionService, public chatService: ChatService) {
         // Get the initial messages from the server
         this.chatService.get_models().subscribe(models => {
             this.models = models;
@@ -111,8 +119,12 @@ export class ChatPageComponent implements OnInit, OnDestroy {
                 next: (chunk) => {
                     if (chunk.type == "text") {
                         this.chatView.addAnswerChunk(chunk.value);
+                        if (this.isSpeech)
+                            this.speechSynthesis.addChunk(chunk.value);
                     } else if (chunk.type.startsWith("error")) {
                         this.chatView.stopTyping()
+                        this.speechSynthesis.stopSpeaking();
+                        this.isSpeech = false;
                         if (this.chatService.chat.history[this.chatService.chat.history.length - 1].author == "ai")
                             this.chatService.chat.history.pop();
                         const errorClass = chunk.type.split(":")[1];
@@ -122,6 +134,8 @@ export class ChatPageComponent implements OnInit, OnDestroy {
                     }
                 },
                 complete: () => {
+                    this.speechSynthesis.endChunks();
+                    this.isSpeech = false;
                     this.chatService.waitingForResponse = false;
                 }
             });
@@ -137,7 +151,10 @@ export class ChatPageComponent implements OnInit, OnDestroy {
         this.chatService.get(chat_session_id).subscribe(() => {
             if (this.isHandset)
                 this.drawerContainer.close();
-            setTimeout(() => this.chatView.scrollBottom(), 100);
+            setTimeout(() => {
+                this.chatView.scrollBottom();
+                this.messageInput.nativeElement.focus()
+            }, 100);
             this.progressSpinner = false;
             this.newMessage = '';
             this.sessionService.clearFiles()
@@ -150,16 +167,17 @@ export class ChatPageComponent implements OnInit, OnDestroy {
         if (this.chatService.waitingForResponse) {
             // Cancel the current request
             this.dataSubscription.unsubscribe();
+            this.chatService.waitingForResponse = false;
             if (this.chatService.chat.history[this.chatService.chat.history.length - 1].author == "ai")
                 this.chatService.chat.history.pop();
             const question = this.chatService.chat.history.pop();
             if (question)
                 this.newMessage = question.content ?? '';
             this.chatService.putChatSession(this.chatService.chat).subscribe();
-        } else {
-            // Stop typing
-            this.chatView.stopTyping();
         }
+        this.chatView.stopTyping();
+        this.speechSynthesis.stopSpeaking();
+        this.isSpeech = false;
     }
 
     deleteChat(chat_session_id: string) {
@@ -183,4 +201,7 @@ export class ChatPageComponent implements OnInit, OnDestroy {
         return this.chatService.isLoading || this.chatService.waitingForResponse || this.chatService.isTyping;
     }
 
+    stopTyping() {
+        setTimeout(() => this.messageInput.nativeElement.focus(), 100);
+    }
 }